修复多次调用IConfigAdapter
