+++ Apollo/Internals/AbstractConfigRepository.cs
@@ -14,28 +14,12 @@ namespace Com.Ctrip.Framework.Apollo.Internals
 
         private readonly List<IRepositoryChangeListener> _listeners = new List<IRepositoryChangeListener>();
         public string Namespace { get; }
-        public ConfigFileFormat Format { get; } = ConfigFileFormat.Properties;
 
-        protected AbstractConfigRepository(string @namespace)
-        {
-            Namespace = @namespace;
+        protected AbstractConfigRepository(string @namespace) => Namespace = @namespace;
 
-            var ext = Path.GetExtension(@namespace);
-            if (ext.Length > 1 && Enum.TryParse(ext.Substring(1), true, out ConfigFileFormat format)) Format = format;
-        }
-
-        public abstract Properties GetRawConfig();
+        public abstract Properties GetConfig();
 
         public abstract Task Initialize();
-
-        public Properties GetConfig()
-        {
-            var properties = GetRawConfig();
-
-            return Format != ConfigFileFormat.Properties && ConfigAdapterRegister.TryGetAdapter(Format, out var adapter)
-                ? adapter.GetProperties(properties)
-                : properties;
-        }
 
         public void AddChangeListener(IRepositoryChangeListener listener)
         {
